@@ -1,4 +1,4 @@
-class Post < ActiveRecord::Base
+class Post < ApplicationRecord
 mount_uploader :image, ImageUploader
 paginates_per 5
 
